Rap/requirements.txt
av==9.2.0
decord==0.6.0
einops==0.4.1
fairscale==0.4.4
ipdb==0.13.9
lmdb==1.2.1
numpy==1.22.3
opencv_python==4.5.5.64
pandas==1.4.1
Pillow==9.2.0
pycocoevalcap==1.2
pycocotools==2.0.4
ruamel.base==1.0.0
tensorboardX==2.5.1
timm==0.4.12
torch==1.11.0
torchvision==0.12.0
transformers==4.15.0
ujson==5.2.0
